[Redux] Move curriculum to separate file
