+++ services/frontend/src/redux/curriculum.ts
@@ -1,0 +1,14 @@
+import { RTKApi } from '@/apiConnection'
+
+const curriculumsApi = RTKApi.injectEndpoints({
+  endpoints: builder => ({
+    getCurriculumOptions: builder.query({
+      query: ({ code }) => `/v3/curriculum-options/${code}`,
+    }),
+    getCurriculums: builder.query({
+      query: ({ code, periodIds }) => `/v3/curriculum/${code}/${periodIds.join(',')}`,
+    }),
+  }),
+})
+
+export const { useGetCurriculumOptionsQuery, useGetCurriculumsQuery } = curriculumsApi
