+++ services/frontend/src/redux/populationCourses.js
@@ -13,16 +13,3 @@ const courseStatisticsApi = RTKApi.injectEndpoints({
 })
 
 export const { useGetStudentListCourseStatisticsQuery } = courseStatisticsApi
-
-const curriculumsApi = RTKApi.injectEndpoints({
-  endpoints: builder => ({
-    getCurriculumOptions: builder.query({
-      query: ({ code }) => `/v3/curriculum-options/${code}`,
-    }),
-    getCurriculums: builder.query({
-      query: ({ code, periodIds }) => `/v3/curriculum/${code}/${periodIds.join(',')}`,
-    }),
-  }),
-})
-
-export const { useGetCurriculumOptionsQuery, useGetCurriculumsQuery } = curriculumsApi
